Update
add: stamina
add: settings and menu
chg: main hero attack animation stopper

+++ CURSED/Assets/Scripts/AI/CharacterStats.cs
@@ -9,6 +9,6 @@ public class CharacterStats : MonoBehaviour
     public int currentHealth;
 
     public int staminaLevel = 10;
-    public int maxStamina;
-    public int currentStamina;
+    public float maxStamina;
+    public float currentStamina;
 }
